Extracted the MovieDetails component logic into a new useMovieDetails custom hook
